+++ test/src/utils/ignore_cancellation_test.dart
@@ -26,25 +26,52 @@ void main() {
     expect(completed, isFalse);
   });
 
+  test('rethrows operation exceptions if onError is null', () {
+    expect(
+      ignoreCancellation(() => Future.error(_OperationException())),
+      throwsA(isA<_OperationException>()),
+    );
+  });
+
+  test('rethrows other exceptions if onError is null', () {
+    expect(
+      ignoreCancellation(() => Future.error(_OperationException())),
+      throwsA(isA<_OperationException>()),
+    );
+  });
+
+  test('doesn\'t rethrow operation exceptions if onError is not null', () {
+    expect(
+      ignoreCancellation(
+        () => Future.error(_OperationException()),
+        onError: (e, stackTrace) {},
+      ),
+      completes,
+    );
+  });
+
+  test('rethrows onError exceptions', () {
+    expect(
+      ignoreCancellation(
+        () => Future.error(_OperationException()),
+        onError: (e, stackTrace) => throw _OnErrorException(),
+      ),
+      throwsA(isA<_OnErrorException>()),
+    );
+  });
+
   group('onError', () {
-    test('rethrows other exceptions if onError is null', () {
-      expect(
-        ignoreCancellation(() => Future.error(_TestException())),
-        throwsA(isA<_TestException>()),
+    test('is called if operation throws an exception', () async {
+      Object? exception;
+      await ignoreCancellation(
+        () => Future.error(_OperationException()),
+        onError: (e, stackTrace) => exception = e,
       );
+
+      expect(exception, isNotNull);
     });
 
-    test('doesn\'t rethrow other exceptions if onError is not null', () {
-      expect(
-        ignoreCancellation(
-          () => Future.error(_TestException()),
-          onError: (e, stackTrace) {},
-        ),
-        completes,
-      );
-    });
-
-    test('doesn\'t call onError for CancelledExceptions', () async {
+    test('is not called if operation throws CancelledExceptions', () async {
       final CancellationToken token = CancellationToken()..cancel();
 
       Object? exception;
@@ -55,20 +82,10 @@ void main() {
 
       expect(exception, isNull);
     });
-
-    test('calls onError if other exceptions are thrown', () async {
-      Object? exception;
-      await ignoreCancellation(
-        () => Future.error(_TestException()),
-        onError: (e, stackTrace) => exception = e,
-      );
-
-      expect(exception, isNotNull);
-    });
   });
 
   group('whenComplete', () {
-    test('is called when the operation completes successfully', () async {
+    test('is called when operation completes successfully', () async {
       bool called = false;
       await ignoreCancellation(
         () => Future.value(),
@@ -78,18 +95,47 @@ void main() {
       expect(called, isTrue);
     });
 
-    test('is called when the operation throws an exception', () async {
+    test('is called when operation throws an exception and onError is null',
+        () async {
+      bool called = false;
+      await expectLater(
+        ignoreCancellation(
+          () => Future.error(_OperationException()),
+          whenComplete: () => called = true,
+        ),
+        throwsA(isA<_OperationException>()),
+      );
+
+      expect(called, isTrue);
+    });
+
+    test('is called when operation throws an exception and onError is not null',
+        () async {
       bool called = false;
       await ignoreCancellation(
-        () => Future.error(_TestException()),
-        onError: (e, stackTrace) => {},
+        () => Future.error(_OperationException()),
+        onError: (e, stackTrace) {},
         whenComplete: () => called = true,
       );
 
       expect(called, isTrue);
     });
 
-    test('is not called when the operation is cancelled', () async {
+    test('is called when onError throws an exception', () async {
+      bool called = false;
+      await expectLater(
+        ignoreCancellation(
+          () => Future.error(_OperationException()),
+          onError: (e, stackTrace) => throw _OnErrorException(),
+          whenComplete: () => called = true,
+        ),
+        throwsA(isA<_OnErrorException>()),
+      );
+
+      expect(called, isTrue);
+    });
+
+    test('is not called when operation is cancelled', () async {
       final CancellationToken token = CancellationToken()..cancel();
 
       bool called = false;
@@ -103,7 +149,7 @@ void main() {
   });
 
   group('whenCompleteOrCancelled', () {
-    test('is called when the operation completes successfully', () async {
+    test('is called when operation completes successfully', () async {
       bool called = false;
       await ignoreCancellation(
         () => Future.value(),
@@ -113,18 +159,47 @@ void main() {
       expect(called, isTrue);
     });
 
-    test('is called when the operation throws an exception', () async {
+    test('is called when operation throws an exception and onError is null',
+        () async {
+      bool called = false;
+      await expectLater(
+        ignoreCancellation(
+          () => Future.error(_OperationException()),
+          whenCompleteOrCancelled: () => called = true,
+        ),
+        throwsA(isA<_OperationException>()),
+      );
+
+      expect(called, isTrue);
+    });
+
+    test('is called when operation throws an exception and onError is not null',
+        () async {
       bool called = false;
       await ignoreCancellation(
-        () => Future.error(_TestException()),
-        onError: (e, stackTrace) => {},
+        () => Future.error(_OperationException()),
+        onError: (e, stackTrace) {},
         whenCompleteOrCancelled: () => called = true,
       );
 
       expect(called, isTrue);
     });
 
-    test('is not called when the operation is cancelled', () async {
+    test('is called when onError throws an exception', () async {
+      bool called = false;
+      await expectLater(
+        ignoreCancellation(
+          () => Future.error(_OperationException()),
+          onError: (e, stackTrace) => throw _OnErrorException(),
+          whenCompleteOrCancelled: () => called = true,
+        ),
+        throwsA(isA<_OnErrorException>()),
+      );
+
+      expect(called, isTrue);
+    });
+
+    test('is not called when operation is cancelled', () async {
       final CancellationToken token = CancellationToken()..cancel();
 
       bool called = false;
@@ -138,4 +213,6 @@ void main() {
   });
 }
 
-class _TestException implements Exception {}
+class _OperationException implements Exception {}
+
+class _OnErrorException implements Exception {}